Everywhere: Properly align image bases
Previously, we only aligned the maximum address, which practically never
leads to us getting an aligned address. To resolve this, a basic ranged
and aligned page allocator has been added. This patch also makes it so
that images always get relocated to a new memory address from the file
base, since the images are not initially loaded to an aligned memory
address.